move cssengine dependency to package
