+++ packages/project/src/shared/styles/index.ts
@@ -1,0 +1,1 @@
+export * from "./global-rules";
